requirements.txt
beautifulsoup4==4.12.3
Flask==3.1.0
nltk==3.2.5
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
PyYAML==6.0.2